Making a side menu
Now on the left, there's a menu to see what teams you have selected. In the future, this will also include types of data the user wants to see. Functionality has yet to be added to team buttons. More resizing and messing with the scss files is needed to make it cohesive with the team components