Preparing for added features of searchbar
I want it so that when the user clicks out of the searchbar area, the results disappear. I want to get experience using react hooks to do this, so I'm setting it up for now by getting the active and hidden classes ready, but added functionality will come in future.