Allowing scrolling if button pressed
Finds the id of a team component, then scrolls to it. The function I used, scroll into view is largely supported. However, on most notably, Safari and samsung internet, options are not supported. Therefore, it's a harsh cut to the element. I suggest you switch to chrome or any other browser lol. hopefully this isn't a terrible problem during comp