Allowing user information
user can now input data that they wish, accepts a google sheet link and csv files. More work will probably have to be done in order to style it, but it actually allows this project to be useful. The end is in sight! (I know I said this a month ago, but it's even closer now)